+++ src/geos-unary-atomic.c
@@ -29,7 +29,7 @@
     int resultCode = _func(handle, geometry, &pResult[i]);     \
                                                                \
     if (resultCode == _errorValue) {                           \
-      Rf_error("[%d] %s", i + 1, globalErrorMessage);                            \
+      Rf_error("[%ld] %s", (long)i + 1, globalErrorMessage);                            \
     }                                                          \
   }                                                            \
                                                                \
@@ -131,7 +131,7 @@ SEXP geos_c_extent(SEXP geom) {
 
     // This gets fired on EMPTY, but we catch this case above
     if (resultCode == 0) {
-      Rf_error("[%d] %s", i + 1, globalErrorMessage);
+      Rf_error("[%ld] %s", (long)i + 1, globalErrorMessage);
     }
   }
 
@@ -188,7 +188,7 @@ SEXP geos_c_hilbert_code(SEXP geom, SEXP geomExtent, SEXP level_sexp) {
     );
 
     if (resultCode != 1) {
-      Rf_error("[%d] %s", i + 1, globalErrorMessage);
+      Rf_error("[%ld] %s", (long)i + 1, globalErrorMessage);
     }
 
     code[i] = item_code;
@@ -226,7 +226,7 @@ SEXP geos_c_hilbert_code(SEXP geom, SEXP geomExtent, SEXP level_sexp) {
     int resultCode = _func(handle, geometry);                     \
                                                                   \
     if (resultCode == _errorValue) {                              \
-      Rf_error("[%d] %s", i + 1, globalErrorMessage);                               \
+      Rf_error("[%ld] %s", (long)i + 1, globalErrorMessage);                               \
     } else {                                                      \
       pResult[i] = resultCode;                                    \
     }                                                             \
@@ -323,14 +323,14 @@ SEXP geos_c_is_clockwise(SEXP geom) {
     seq = GEOSGeom_getCoordSeq_r(handle, geometry);
     // e.g., when not a point, linestring, or linearring
     if (seq == NULL) {
-      Rf_error("[%d] %s", i + 1, globalErrorMessage);
+      Rf_error("[%ld] %s", (long)i + 1, globalErrorMessage);
     }
 
     int resultCode = GEOSCoordSeq_isCCW_r(handle, seq, &isCCW);
 
     // e.g., not enough points in ring
     if (resultCode == 0) {
-      Rf_error("[%d] %s", i + 1, globalErrorMessage);
+      Rf_error("[%ld] %s", (long)i + 1, globalErrorMessage);
     }
 
     pResult[i] = !isCCW;
@@ -384,7 +384,7 @@ SEXP geos_c_is_valid_detail(SEXP geom, SEXP allowSelfTouchingRingFormingHole) {
     // error
     // (don't know how to trigger this error)
     if (validResult == 2) {
-      Rf_error("[%d] %s", i + 1, globalErrorMessage); // # nocov
+      Rf_error("[%ld] %s", (long)i + 1, globalErrorMessage); // # nocov
     } else if (validResult == 1) {
       pResultIsValid[i] = 1;
       SET_STRING_ELT(resultReason, i, NA_STRING);
