+++ src/geos-unary-geometry.c
@@ -27,7 +27,7 @@
     geometryResult = _func(handle, geometry);                  \
                                                                \
     if (geometryResult == NULL) {                              \
-      Rf_error("[%d] %s", i + 1, globalErrorMessage);                            \
+      Rf_error("[%ld] %s", (long)i + 1, globalErrorMessage);                            \
     } else {                                                   \
       SET_VECTOR_ELT(result, i, geos_common_geometry_xptr(geometryResult));\
     }                                                          \
@@ -201,7 +201,7 @@ SEXP geos_c_transform_xy(SEXP geom, SEXP trans_xptr) {
     geometryResult = GEOSGeom_transformXY_r(handle, geometry, &transform_callback, &data);
 
     if (geometryResult == NULL) {
-      Rf_error("[%d] %s", i + 1, globalErrorMessage);
+      Rf_error("[%ld] %s", (long)i + 1, globalErrorMessage);
     }
 
     SET_VECTOR_ELT(result, i, geos_common_geometry_xptr(geometryResult));
@@ -271,7 +271,7 @@ SEXP geos_c_make_valid_with_params(SEXP geom, SEXP params_sexp) {
 
     if (geometryResult == NULL) {
       GEOSMakeValidParams_destroy_r(handle, params);
-      Rf_error("[%d] %s", i + 1, globalErrorMessage);
+      Rf_error("[%ld] %s", (long)i + 1, globalErrorMessage);
     } else {
       SET_VECTOR_ELT(result, i, geos_common_geometry_xptr(geometryResult));
     }
@@ -310,7 +310,7 @@ SEXP geos_c_make_valid_with_params(SEXP geom, SEXP params_sexp) {
     geometryResult = _call;                                                \
                                                                            \
     if (geometryResult == NULL) {                                          \
-      Rf_error("[%d] %s", i + 1, globalErrorMessage);                                        \
+      Rf_error("[%ld] %s", (long)i + 1, globalErrorMessage);                                        \
     } else {                                                               \
       SET_VECTOR_ELT(result, i, geos_common_geometry_xptr(geometryResult));\
     }                                                                      \
@@ -474,7 +474,7 @@ SEXP geos_c_set_srid(SEXP geom, SEXP srid) {
 
     // don't know how to trigger this
     if (geometryResult == NULL) {
-      Rf_error("[%d] %s", i + 1, globalErrorMessage); // # nocov
+      Rf_error("[%ld] %s", (long)i + 1, globalErrorMessage); // # nocov
     }
 
     // has no return code for exception
@@ -515,14 +515,14 @@ SEXP geos_c_normalize(SEXP geom) {
 
     // don't know how to trigger this
     if (geometryResult == NULL) {
-      Rf_error("[%d] %s", i + 1, globalErrorMessage); // # nocov
+      Rf_error("[%ld] %s", (long)i + 1, globalErrorMessage); // # nocov
     }
 
     returnCode = GEOSNormalize_r(handle, geometryResult);
 
     if (returnCode == -1) {
       GEOSGeom_destroy_r(handle, geometryResult); // # nocov
-      Rf_error("[%d] %s", i + 1, globalErrorMessage); // # nocov
+      Rf_error("[%ld] %s", (long)i + 1, globalErrorMessage); // # nocov
     }
 
     SET_VECTOR_ELT(result, i, geos_common_geometry_xptr(geometryResult));
@@ -568,7 +568,7 @@ SEXP geos_c_minimum_bounding_circle(SEXP geom) {
     geometryResult = GEOSMinimumBoundingCircle_r(handle, geometry, pRadius + i, &center);
 
     if (geometryResult == NULL) {
-      Rf_error("[%d] %s", i + 1, globalErrorMessage);
+      Rf_error("[%ld] %s", (long)i + 1, globalErrorMessage);
     } else {
       SET_VECTOR_ELT(result, i, geos_common_geometry_xptr(geometryResult));
       GEOSGeomGetX_r(handle, center, pX + i);
@@ -612,7 +612,7 @@ SEXP geos_c_clip_by_rect(SEXP geom, SEXP xmin, SEXP ymin, SEXP xmax, SEXP ymax) 
     geometryResult = GEOSClipByRect_r(handle, geometry, pXmin[i], pYmin[i], pXmax[i], pYmax[i]);
 
     if (geometryResult == NULL) {
-      Rf_error("[%d] %s", i + 1, globalErrorMessage);
+      Rf_error("[%ld] %s", (long)i + 1, globalErrorMessage);
     } else {
       SET_VECTOR_ELT(result, i, geos_common_geometry_xptr(geometryResult));
     }
@@ -649,7 +649,7 @@ SEXP geos_c_delaunay_triangulation(SEXP geom, SEXP tolerace, SEXP edges) {
     geometryResult = GEOSDelaunayTriangulation_r(handle, geometry, dTolerance, iEdges);
 
     if (geometryResult == NULL) {
-      Rf_error("[%d] %s", i + 1, globalErrorMessage);
+      Rf_error("[%ld] %s", (long)i + 1, globalErrorMessage);
     } else {
       SET_VECTOR_ELT(result, i, geos_common_geometry_xptr(geometryResult));
     }
@@ -694,7 +694,7 @@ SEXP geos_c_voronoi_diagram(SEXP geom, SEXP env, SEXP tolerace, SEXP edges) {
     geometryResult = GEOSVoronoiDiagram_r(handle, geometry, envGeometry, dTolerance, iEdges);
 
     if (geometryResult == NULL) {
-      Rf_error("[%d] %s", i + 1, globalErrorMessage);
+      Rf_error("[%ld] %s", (long)i + 1, globalErrorMessage);
     } else {
       SET_VECTOR_ELT(result, i, geos_common_geometry_xptr(geometryResult));
     }
@@ -763,7 +763,7 @@ SEXP geos_c_voronoi_diagram(SEXP geom, SEXP env, SEXP tolerace, SEXP edges) {
                                                                  \
     if (geometryResult == NULL) {                                \
       GEOSBufferParams_destroy_r(handle, bufferParams);          \
-      Rf_error("[%d] %s", i + 1, globalErrorMessage);                              \
+      Rf_error("[%ld] %s", (long)i + 1, globalErrorMessage);                              \
     } else {                                                     \
       SET_VECTOR_ELT(result, i, geos_common_geometry_xptr(geometryResult));\
     }                                                            \
@@ -821,7 +821,7 @@ SEXP geos_c_geometry_n(SEXP geom, SEXP n) {
 
     // don't know how to make this occur
     if (geometryResult == NULL) {
-      Rf_error("[%d] %s", i + 1, globalErrorMessage); // # nocov
+      Rf_error("[%ld] %s", (long)i + 1, globalErrorMessage); // # nocov
     }
 
     SET_VECTOR_ELT(result, i, geos_common_child_geometry_xptr(geometryResult, item));
@@ -880,7 +880,7 @@ SEXP geos_c_ring_n(SEXP geom, SEXP n) {
 
     // don't know how to make this occur
     if (geometryResult == NULL) {
-      Rf_error("[%d] %s", i + 1, globalErrorMessage); // # nocov
+      Rf_error("[%ld] %s", (long)i + 1, globalErrorMessage); // # nocov
     }
 
     SET_VECTOR_ELT(result, i, geos_common_child_geometry_xptr(geometryResult, item));
